rename pong to ping and add ping tests
