+++ src/modules/commands/ping.js
